simple_app/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
PyYAML==6.0.1
Markdown==3.5.1
